// dops/src/enum/notification-template.enum.ts
export enum NotificationTemplate {
  PROFILE_REGISTRATION = 'PROFILE_REGISTRATION',
  ISSUE_PERMIT = 'ISSUE_PERMIT',
  PAYMENT_RECEIPT = 'PAYMENT_RECEIPT',
  COMPANY_SUSPEND = 'COMPANY_SUSPEND',
  COMPANY_UNSUSPEND = 'COMPANY_UNSUSPEND',
  APPLICATION_APPROVED = 'APPLICATION_APPROVED',
  APPLICATION_REJECTED = 'APPLICATION_REJECTED',
}